完成注册用户的编码实现测试

+++ ExamApp/ExamApp/RegisterData.h
@@ -23,5 +23,5 @@
 //手机号码
 @property(nonatomic,copy)NSString *phone;
 //根据Tag值设置值
--(void)setValue:(NSString *)value forTag:(int)tag;
+-(void)setValue:(NSString *)value forTag:(NSInteger)tag;
 @end
